Move test-tree-node.ts to /query-testing/test-tree-node.ts
