+++ extensions/ql-vscode/src/test-ui.ts
@@ -7,7 +7,7 @@ import {
   TestEvent,
   TestSuiteEvent,
 } from "vscode-test-adapter-api";
-import { TestTreeNode } from "./test-tree-node";
+import { TestTreeNode } from "./query-testing/test-tree-node";
 import { DisposableObject } from "./pure/disposable-object";
 import { QLTestAdapter } from "./query-testing/test-adapter";
 import { App } from "./common/app";
